Rename Georgia_De Kalb County.html to Georgia_DeKalb County.html
The file name had a space in the county name (it is spelled DeKalb, not De Kalb). This was breaking the link in the main page. I'm not sure if this change also needs to be in the code that generates the files, so that would need to be checked. Thanks.